scope.api/app/requirements.txt
django
mysqlclient
djangorestframework
markdown
django-filter
coreapi
pygments
markdown
raven